% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.handleUniqueDim}
\alias{.handleUniqueDim}
\title{Handle unique haplotype dimension issue}
\usage{
.handleUniqueDim(unique_haplo)
}
\arguments{
\item{unique_haplo}{Unique haplotype result}
}
\description{
Convert unique result to matrix if dimension degenerates
}
\keyword{internal}
