% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.safeDivide}
\alias{.safeDivide}
\title{Safe division}
\usage{
.safeDivide(numerator, denominator)
}
\arguments{
\item{numerator}{Numerator}

\item{denominator}{Denominator}
}
\description{
Division with protection against zero denominator
}
\keyword{internal}
