% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.sortUniqueHaplotypes}
\alias{.sortUniqueHaplotypes}
\title{Sort unique haplotypes}
\usage{
.sortUniqueHaplotypes(unique_haplo)
}
\arguments{
\item{unique_haplo}{Unique haplotype matrix}
}
\description{
Sort unique haplotypes by site values
}
\keyword{internal}
