% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.validateInput}
\alias{.validateInput}
\title{Validate input data}
\usage{
.validateInput(x, name = "input", check_values = FALSE)
}
\arguments{
\item{x}{Input object}

\item{name}{Object name for error message}

\item{check_values}{Whether to check for 0/1 values (default: FALSE)}
}
\value{
Matrix representation of input
}
\description{
Check input matrix validity, enforce matrix type, and validate values
}
\keyword{internal}
