% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.checkParallelAvailable}
\alias{.checkParallelAvailable}
\title{Check parallel package availability}
\usage{
.checkParallelAvailable(parallel, n_workers)
}
\description{
Check if parallel processing is available and valid
}
